Refactor on models
label: technical debt

We have many blueprints, so definition of database structure should be
in upper directory.

Resolves #xx
